rename onboarding, splash, and qpi sem, cleanup
