+++ lib/screens/splash_page.dart
@@ -3,7 +3,7 @@ import 'dart:async';
 import 'package:flutter/material.dart';
 
 import 'package:admu_student_app/main.dart';
-import 'package:admu_student_app/screens/login.dart';
+import 'package:admu_student_app/screens/login_page.dart';
 
 class SplashPage extends StatefulWidget {
   @override
@@ -31,7 +31,7 @@ class StartState extends State<SplashPage> {
     Navigator.pushReplacement(
         context,
         MaterialPageRoute(
-          builder: (context) => Login(),
+          builder: (context) => LoginPage(),
         ));
   }
