+++ lib/widgets/qpi/qpi_view.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:google_fonts/google_fonts.dart';
 
 import 'package:admu_student_app/main.dart';
